Suppress MSVC warning about "partial" formal argument being unused.
When using large keys, the algorithm to search a bucket will initially try to
compare against just the first byte of the key, rather than comparing the entire
thing. But if the key is "simple" (POD, less than 8 bytes), this optimization is
skipped. In the case of the simple key, MSVC optimizes away the use of the
"partial" key, which is great, but then it warns that the passed in variable was
unused. This commit adds cast-to-void to the impacted methods in order to
silence the warning.