Address some 'did you want a 64-bit shift?' type warnings from MSVC 2015.
1>cuckoohash_map.hh(1814): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
1>libcuckoo_bucket_container.hh(189): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
1>cuckoohash_map.hh(1591): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)

There are all places where we shift an unsigned long (1UL) by a size_type. When
doing a 64-bit build MSVC 2015 warns about this having a 32-bit result - on
Windows "long" is only 32-bits. This commit fixes this issue by explicitly
specifying the value to be shifted as size_type.