GCC 4.7 build breakage due to missing std::is_nothrow_destructible().
GCC 4.7 fails to build with a handful of errors like:

  libcuckoo_bucket_container.hh: In member function ‘void libcuckoo_bucket_container<Key, T, Allocator, Partial, SLOT_PER_BUCKET>::clear()’:
  libcuckoo_bucket_container.hh:222:9: error: ‘is_nothrow_destructible’ is not a member of ‘std’
  libcuckoo_bucket_container.hh:222:46: error: expected primary-expression before ‘>’ token
  libcuckoo_bucket_container.hh:222:47: error: ‘::value’ has not been declared

This is because GCC 4.7 does not have an implementation of
std::is_nothrow_destructible(). As of this commit libcuckoo_bucket_container now
contains its own implementation of is_nothrow_destructible() on GCC 4.7, which
delegates to std::is_nothrow_destructible() everywhere else.